feat: absorb Linux & BSD distro innovations and enhance Shell REPL CLI
- Absorb OpenBSD pledge/unveil, FreeBSD jail, Void runit, Gentoo USE-flags, NixOS declarative manager, and Arch AUR package management innovations into core kernel and package layers.
- Expose interactive CLI REPL commands (`pledge`, `jail`, `runit`, `useflags`, `nixos`, `aur`) in `src/shell/repl.rs`.
- Fix capability bitmask logic, type mismatches, and duplicate definitions across `src/kernel/component.rs`, `src/security/capability.rs`, `src/filesystem/vfs.rs`, `src/compatibility/fedora.rs`, and `src/power/mod.rs`.
- Ensure 100% clean compilation and passing unit and integration test suite across all targets.

+++ src/filesystem/support.rs
@@ -60,7 +60,7 @@ impl Filesystem for SimpleFilesystem {
         self.id
     }
     fn fs_type(&self) -> FilesystemType {
-        let val = self.fs_type.load(Ordering::SeqCst) as u32; unsafe { core::mem::transmute(val) }
+        let val = self.fs_type.load(Ordering::SeqCst); unsafe { core::mem::transmute(val) }
     }
 
     fn mount(&mut self, _device: &[u8], mountpoint: &[u8]) -> Result<(), FilesystemError> {
